refactor: :recycle: разорвать цикл профиля xml-синхронизации
Вынести добавление путей BaseForm из componentProfile в отдельный модуль плана. Устранить взаимную зависимость componentProfile и типов full sync без изменения поведения.
