+++ packages/core/metadata/fullSyncToXml/componentProfile.ts
@@ -5,7 +5,6 @@ import type {
 } from "../configurationIndex"
 import type { ConfirmedComponentState } from "../project/componentState/types"
 import type { XMLDefaultVariant } from "../context/types"
-import type { FullXmlSyncPlan } from "./types"
 
 export type XmlSyncProfileKind = "configuration" | "configurationExtension"
 
@@ -49,32 +48,6 @@ export interface FullXmlSyncComponentProfile {
     readonly runtime: FullXmlSyncProfileRuntime
     readonly rootYaml: unknown
   }) => FullXmlSyncProfileRuntime
-}
-
-export function attachBorrowedFormPaths(
-  plan: FullXmlSyncPlan,
-  runtime: Pick<FullXmlSyncProfileRuntime, "borrowedForms">,
-): FullXmlSyncPlan {
-  const borrowedByAddress = new Map(
-    (runtime.borrowedForms ?? []).map((form) => [form.logicalAddress, form])
-  )
-  return {
-    ...plan,
-    assignments: plan.assignments.map((assignment) => {
-      const borrowed = borrowedByAddress.get(assignment.logicalAddress)
-      return borrowed === undefined
-        ? assignment
-        : {
-            ...assignment,
-            baseFormPaths: {
-              baseProjectPath: borrowed.baseProjectPath,
-              ...(borrowed.savedProjectPath === undefined
-                ? {}
-                : { savedProjectPath: borrowed.savedProjectPath }),
-            },
-          }
-    }),
-  }
 }
 
 const profiles = new Map<XmlSyncProfileKind, FullXmlSyncComponentProfile>()
