+++ packages/core/metadata/fullSyncToXml/syncConfiguration.ts
@@ -17,7 +17,8 @@ import {
 import { createValidationProfiler } from "../validation/profile"
 import type { Diagnostic } from "../validation/types"
 import type { ProjectStateComponentProjection, ProjectStateReadToken, ProjectStateService } from "../projectState"
-import { attachBorrowedFormPaths, resolveFullXmlSyncComponentProfile } from "./componentProfile"
+import { resolveFullXmlSyncComponentProfile } from "./componentProfile"
+import { attachBorrowedFormPaths } from "./borrowedFormPlan"
 import { buildXmlSyncPlan, type XmlSyncSelection } from "./selection"
 import { createFullXmlSyncCompositionSnapshot } from "./sharedMetadata"
 import { transferFullXmlSyncExternalFiles } from "./transferExternalFiles"
